pyproject.toml
# File Purpose:
# - Define Python project metadata and enforce static quality/security tooling baselines.
#
# Key Security Considerations:
# - Strict linting and typing reduce unsafe patterns and runtime ambiguity.
#
# OWASP 2025 Categories Addressed:
# - A03, A05, A10

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-ci"
version = "0.1.0"
description = "Security automation toolkit for GitHub workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.110,<0.120",
	"pydantic>=2.6,<3.0",
	"supabase>=2.4,<3.0",
	"python-jose>=3.3,<4.0",
	"argon2-cffi>=23.1,<24.0",
	"httpx>=0.27,<0.28",
	"weasyprint>=60,<61",
]

[project.optional-dependencies]
dev = [
	"ruff>=0.11,<0.12",
	"mypy>=1.15,<1.16",
	"pytest>=9.0,<10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "N", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"api/utils/crypto.py" = ["I001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
strict_equality = true
pretty = true
mypy_path = "."
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["api*"]
exclude = ["supabase*", "terraform*", "tests*"]
